requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
python-dotenv>=1.0.1

# ML / Data
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0

# LLM Providers (install the ones you need)
anthropic>=0.28.0
openai>=1.30.0

# Log Sources (install the ones you need)
boto3>=1.34.0          # AWS CloudWatch
datadog-api-client>=2.24.0  # Datadog

# Utilities
httpx>=0.27.0
pydantic>=2.7.0
rich>=13.7.0           # Pretty console output
